@@ -20,62 +25,63 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.0",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/forms": "~13.3.0",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
-    "rxjs": "~7.4.0",
+    "@angular/animations": "15.0.4",
+    "@angular/common": "15.0.4",
+    "@angular/compiler": "15.0.4",
+    "@angular/core": "15.0.4",
+    "@angular/forms": "15.0.4",
+    "@angular/platform-browser": "15.0.4",
+    "@angular/platform-browser-dynamic": "15.0.4",
+    "@angular/router": "15.0.4",
+    "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.0",
-    "@angular-eslint/eslint-plugin": "~13.1.0",
-    "@angular-eslint/eslint-plugin-template": "~13.1.0",
-    "@angular-eslint/template-parser": "~13.1.0",
-    "@angular/cli": "~13.3.0",
-    "@angular/compiler-cli": "~13.3.0",
-    "@angular/language-service": "~13.3.0",
+    "@angular-devkit/build-angular": "15.0.4",
+    "@angular-eslint/eslint-plugin": "15.0.0",
+    "@angular-eslint/eslint-plugin-template": "15.0.0",
+    "@angular-eslint/template-parser": "15.0.0",
+    "@angular/cli": "~15.0.0",
+    "@angular/compiler-cli": "15.0.4",
+    "@angular/language-service": "15.0.4",
     "@commitlint/cli": "^16.2.4",
     "@commitlint/config-angular": "^16.2.4",
     "@commitlint/config-conventional": "^16.2.4",
     "@hirez_io/jest-given": "^1.1.0",
     "@hirez_io/observer-spy": "^2.2.0",
     "@jscutlery/semver": "^2.23.4",
-    "@nrwl/angular": "^14.1.5",
-    "@nrwl/cli": "14.1.4",
-    "@nrwl/cypress": "14.1.5",
-    "@nrwl/eslint-plugin-nx": "14.1.4",
-    "@nrwl/jest": "14.1.4",
-    "@nrwl/js": "14.1.4",
-    "@nrwl/linter": "14.1.4",
-    "@nrwl/nx-cloud": "latest",
-    "@nrwl/workspace": "14.1.4",
-    "@types/jest": "27.4.1",
-    "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "~5.18.0",
-    "@typescript-eslint/parser": "~5.18.0",
+    "@nrwl/angular": "15.3.3",
+    "@nrwl/cli": "15.3.3",
+    "@nrwl/cypress": "15.3.3",
+    "@nrwl/eslint-plugin-nx": "15.3.3",
+    "@nrwl/jest": "15.3.3",
+    "@nrwl/js": "15.3.3",
+    "@nrwl/linter": "15.3.3",
+    "@nrwl/nx-cloud": "15.0.2",
+    "@nrwl/workspace": "15.3.3",
+    "@types/jest": "28.1.8",
+    "@types/node": "18.11.17",
+    "@typescript-eslint/eslint-plugin": "5.36.1",
+    "@typescript-eslint/parser": "5.36.1",
     "cypress": "^9.1.0",
-    "eslint": "~8.12.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "git-cz": "^4.8.0",
     "husky": "^7.0.4",
-    "jest": "27.5.1",
-    "jest-preset-angular": "11.1.1",
-    "ng-packagr": "^13.0.0",
-    "nx": "14.1.4",
+    "jest": "28.1.3",
+    "jest-environment-jsdom": "28.1.1",
+    "jest-preset-angular": "12.2.3",
+    "ng-packagr": "15.0.3",
+    "nx": "15.3.3",
     "postcss": "^8.4.5",
-    "postcss-import": "^14.0.2",
-    "postcss-preset-env": "^6.7.0",
-    "postcss-url": "^10.1.1",
-    "prettier": "^2.5.1",
-    "ts-jest": "27.1.4",
-    "ts-node": "9.1.1",
-    "typescript": "~4.6.2"
+    "postcss-import": "14.1.0",
+    "postcss-preset-env": "7.5.0",
+    "postcss-url": "10.1.3",
+    "prettier": "2.6.2",
+    "ts-jest": "28.0.8",
+    "ts-node": "10.9.1",
+    "typescript": "4.8.4"
   },
   "config": {
     "commitizen": {
