@@ -1,6 +1,6 @@
 const { getJestProjects } = require('@nrwl/jest');
 
-module.exports = {
+export default {
   projects: getJestProjects(),
   setupFilesAfterEnv: ['node_modules/@hirez_io/jest-given/dist/jest-given.js'],
 };
